fix: fix openapi response, extract formatter into its own file
