+++ packages/common/src/openapi/paths/projectRunQuery.json
@@ -45,9 +45,6 @@
                             },
                             "tableCalculations": {
                                 "type": "array"
-                            },
-                            "filters": {
-                                "type": "array"
                             }
                         },
                         "required": ["metrics", "dimensions"]
