+++ packages/common/src/openapi/schemas/QueryResult.json
@@ -8,7 +8,13 @@
                 "type": "object",
                 "properties": {
                     "raw": {
-                        "type": "number"
+                        "oneOf": [
+                            { "type": "string" },
+                            { "type": "number" },
+                            { "type": "boolean" },
+                            { "type": "array" },
+                            { "type": "object" }
+                        ]
                     },
                     "formatted": {
                         "type": "string"
